get rid of return type for export, since it can bool or string
