+++ src/Umpirsky/ListGenerator/Exporter/ExporterInterface.php
@@ -8,7 +8,7 @@ interface ExporterInterface
      * Exports data into specific format.
      *
      * @param  array  $data
-     * @return string
+     * @return string|false
      */
-    public function export(array $data): string;
+    public function export(array $data);
 }
