+++ src/Umpirsky/ListGenerator/Exporter/Format/Php.php
@@ -9,7 +9,7 @@ class Php extends Exporter
     /**
      * {@inheritdoc}
      */
-    public function export(array $data): string
+    public function export(array $data)
     {
         return sprintf('<?php return %s;%s', var_export($data, true), PHP_EOL);
     }
